auth0-fastapi/pyproject.toml
[tool.poetry]
name = "fastapi-python"
version = "0.1.0"
description = "This package provides base auth capability for Auth0 Fast API."
license = "apache-2.0"
homepage = "https://auth0.com"
authors = [
    "Snehil <snehil.kishorea@okta.com>",
]
readme = "README.md"
packages = [{ include = "auth" }] 

[tool.poetry.dependencies]
python = "^3.6"
auth0_python = "^4.8.0"
cryptography = ">=43.0.1" # pyjwt has a weak dependency on cryptography
pyjwt = ">=2.8.0"
fastapi = {version = "^0.115.0", extras = ["standard"]}

[tool.poetry.group.test.dependencies]
pytest-randomly = "^3.15.0"
pytest-asyncio = "^0.25.0"
pytest = "^8.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"